+++ app/src/main/java/com/tari/android/wallet/model/PendingOutboundTx.kt
@@ -35,6 +35,7 @@ package com.tari.android.wallet.model
 import android.os.Parcel
 import android.os.Parcelable
 import com.tari.android.wallet.ffi.FFICompletedTx
+import com.tari.android.wallet.ffi.FFIPendingOutboundTx
 import java.math.BigInteger
 
 /**
@@ -46,18 +47,19 @@ class PendingOutboundTx() : Tx(), Parcelable {
 
     var fee = MicroTari(BigInteger("0"))
 
-    constructor(id: BigInteger, user: User, amount: MicroTari, fee: MicroTari, timestamp: BigInteger, message: String, status: TxStatus) : this() {
-        this.id = id
-        this.direction = Direction.OUTBOUND
-        this.user = user
-        this.amount = amount
-        this.fee = fee
-        this.timestamp = timestamp
-        this.message = message
-        this.status = status
+    internal constructor(tx: FFICompletedTx) : this() {
+        this.id = tx.getId()
+        this.direction = tx.getDirection()
+        this.user = tx.getUser()
+        this.amount = MicroTari(tx.getAmount())
+        this.fee = MicroTari(tx.getFee())
+        this.timestamp = tx.getTimestamp()
+        this.message = tx.getMessage()
+        this.status = TxStatus.map(tx.getStatus())
+        tx.destroy()
     }
 
-    internal constructor(tx: FFICompletedTx) : this() {
+    internal constructor(tx: FFIPendingOutboundTx) : this() {
         this.id = tx.getId()
         this.direction = tx.getDirection()
         this.user = tx.getUser()
