+++ app/src/main/java/com/tari/android/wallet/ui/fragment/tx/TxListRouter.kt
@@ -1,9 +1,10 @@
 package com.tari.android.wallet.ui.fragment.tx
 
 import com.tari.android.wallet.model.Tx
+import com.tari.android.wallet.model.TxId
 
 interface TxListRouter {
-    fun toTxDetails(tx: Tx)
+    fun toTxDetails(tx: Tx? = null, txId: TxId? = null)
 
     fun toTTLStore()
