+++ app/src/main/res/layout/fragment_tx_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".ui.activity.tx.TxDetailsActivity">
+    tools:context=".ui.fragment.tx.details.TxDetailsFragment">
 
     <RelativeLayout
         android:id="@+id/payment_header_view"
@@ -55,20 +55,41 @@
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="@id/back_view" />
 
-                <com.tari.android.wallet.ui.component.CustomFontTextView
-                    android:id="@+id/date_text_view"
-                    android:layout_width="wrap_content"
+                <androidx.appcompat.widget.LinearLayoutCompat
+                    android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_gravity="center_horizontal"
                     android:layout_marginTop="12dp"
                     android:layout_marginBottom="24dp"
-                    android:textColor="@color/tx_detail_contact_name_label_text"
-                    android:textSize="13sp"
-                    app:customFont="@string/font_medium"
+                    android:layout_marginHorizontal="52dp"
+                    android:orientation="vertical"
                     app:layout_constraintEnd_toEndOf="@id/payment_state_text_view"
                     app:layout_constraintStart_toStartOf="@id/payment_state_text_view"
-                    app:layout_constraintTop_toBottomOf="@id/payment_state_text_view"
-                    tools:text="November 1st 2019 at 2:39 PM" />
+                    app:layout_constraintTop_toBottomOf="@id/payment_state_text_view">
+
+                    <com.tari.android.wallet.ui.component.CustomFontTextView
+                        android:id="@+id/cancellation_reason_view"
+                        android:visibility="gone"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_gravity="center_horizontal"
+                        android:textColor="@color/secondary_text_color"
+                        android:textSize="13sp"
+                        app:customFont="@string/font_medium"
+                        tools:text="November 1st 2019 at 2:39 PM" />
+
+                    <com.tari.android.wallet.ui.component.CustomFontTextView
+                        android:id="@+id/date_text_view"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_gravity="center_horizontal"
+                        android:layout_marginTop="2dp"
+                        android:textColor="@color/secondary_text_color"
+                        android:textSize="13sp"
+                        app:customFont="@string/font_medium"
+                        tools:text="November 1st 2019 at 2:39 PM" />
+
+                </androidx.appcompat.widget.LinearLayoutCompat>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
