+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="light_gray">#3A000000</color>
     <color name="disabled_cta">@color/dark_gray</color>
     <color name="tx_status">#FFEC9C00</color>
+    <color name="secondary_text_color">#817E8A</color>
 
     <!-- common gradient button -->
     <color name="gradient_btn_start">#7300ff</color>
